nats refactoring
