+++ nova/program/__init__.py
@@ -1,5 +1,4 @@
 from nova.program.function import Program, ProgramPreconditions, program
 from nova.program.runner import ProgramRunner
-from nova.program.store import ProgramStore
 
-__all__ = ["ProgramRunner", "program", "ProgramPreconditions", "ProgramStore", "Program"]
+__all__ = ["ProgramRunner", "program", "ProgramPreconditions", "Program"]
